+++ app/uk/gov/hmrc/agentregistration/shared/testdata/TdBase.scala
@@ -35,7 +35,6 @@ import uk.gov.hmrc.agentregistration.shared.lists.SixOrMoreOfficers
 import uk.gov.hmrc.agentregistration.shared.individual.*
 import uk.gov.hmrc.auth.core.retrieve.Credentials
 import uk.gov.hmrc.agentregistration.shared.companieshouse.CompaniesHouseOfficerRole.LlpMember
-import uk.gov.hmrc.agentregistration.shared.risking.PersonReference
 import uk.gov.hmrc.auth.core.ConfidenceLevel
 
 import java.time.*
